onboarding/requirements.txt
slack_bolt>=1.27.0
slack_sdk>=3.39.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
APScheduler>=3.10.0
